fixed errors with renamed file
