+++ benchmarks.py
@@ -15,8 +15,7 @@ sys.path.insert(0, str(TESTS_DIR))
 import grug
 from grug.entity import Entity, ReraisedGameFnError, StackOverflow, TimeLimitExceeded
 from grug.grug_state import GrugFile, GrugRuntimeErrorType, GrugState
-from grug.grug_value import GrugValue
-from grug.parser import Type, PrimitiveType
+from grug.types import GrugValue, Type, PrimitiveType
 from tests.test_grug import (  
     GrugValueUnion,
     GrugValueWorkaround,
